requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==2.2.6
django-allauth==0.40.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==5.1.2
requests==2.31.0
requests-oauthlib==2.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
